requirements.txt
astroid==1.4.9
attrs==17.2.0
Automat==0.6.0
backports.functools-lru-cache==1.3
boto==2.43.0
colorama==0.3.7
configparser==3.5.0
constantly==15.1.0
deepdiff==3.3.0
dj-database-url==0.4.1
Django==1.10
django-cors-headers==1.3.1
django-crispy-forms==1.6.1
django-filter==0.15.3
django-storages==1.5.1
djangorestframework==3.4.0
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.8.0
dynamic-rest==1.5.0
gunicorn==19.6.0
incremental==17.5.0
inflection==0.3.1
isort==4.2.5
jsonpickle==0.9.5
lazy-object-proxy==1.2.2
Markdown==2.6.7
mccabe==0.5.3
Pillow==3.4.2
psycopg2==2.6.2
PyJWT==1.4.2
pylint==1.6.4
six==1.10.0
virtualenv==15.0.3
whitenoise==3.2.2
wrapt==1.10.8
zope.interface==4.4.1
